requirements.txt
aiohttp~=3.7.4.post0
discord.py~=1.7.3
discord-py-slash-command~=3.0.1
dnspython~=2.1.0
motor~=2.5.1
pymongo~=3.12.0
python-dotenv~=0.19.0